add restassured spring boot test

+++ spring-boot2-jwt-security-v5/src/main/java/com/example/jwt/security/v5/repository/OrganizationRepository.java
@@ -3,7 +3,7 @@ package com.example.jwt.security.v5.repository;
 import com.example.jwt.security.v5.model.Organization;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrganizationRepository extends JpaRepository<Organization, Integer> {
+public interface OrganizationRepository extends JpaRepository<Organization, Long> {
 
     Organization findByName(String name);
